requirements.txt
discord.py
requests
python-dotenv
openai
yt-dlp[default] --pre
deep-translator
easy-pil
pynacl
aiohttp